Accepted: didnt check for negative values

+++ Daily Question/2025/January 2025/2025-01-03_2270_Number_of_Ways_to_Split_Array.cpp
@@ -3,11 +3,11 @@ class Solution {
         int waysToSplitArray(vector<int>& nums) {
             int n = nums.size();
             
-            vector<int> prefixLeft;
+            vector<long long int> prefixLeft;
             prefixLeft.push_back(nums[0]);
             for(int i = 1; i < n; i++) prefixLeft.push_back(prefixLeft[i - 1] + nums[i]);
             
-            vector<int> prefixRight;
+            vector<long long int> prefixRight;
             prefixRight.push_back(nums[n - 1]);
             for(int i = n - 2; i >= 0; i--) prefixRight.push_back(prefixRight.back() + nums[i]);
             reverse(prefixRight.begin(), prefixRight.end());
